requirements.txt
flask
wikipedia-api
httpx
python-dotenv
requests
os
